Disabled concurrent GC for PS4
Disabling concurrent GC for PS4 by default because we currently have no way to pause thread execution. If the GC runs on another thread the main thread execution needs to be paused to be able to copy the stack to be used to obtain conservative roots used for GC. If we run the GC on another thread without copying the main thread stack the main thread stack will not be considered when garbage collecting and objects on the main thread stack will be GC'd incorrectly.